+++ forage-android/src/test/java/com/joinforage/forage/android/network/data/CheckBalanceRepositoryTest.kt
@@ -1,6 +1,5 @@
 package com.joinforage.forage.android.network.data
 
-import com.joinforage.forage.android.VaultType
 import com.joinforage.forage.android.core.telemetry.Log
 import com.joinforage.forage.android.fixtures.givenContentId
 import com.joinforage.forage.android.fixtures.givenEncryptionKey
@@ -40,7 +39,7 @@ class CheckBalanceRepositoryTest : MockServerSuite() {
 
         val logger = Log.getSilentInstance()
         repository = MockServiceFactory(
-            mockVaultSubmitter = MockVaultSubmitter(VaultType.VGS_VAULT_TYPE),
+            mockVaultSubmitter = MockVaultSubmitter(),
             mockPinCollector = pinCollector,
             logger = logger,
             server = server
