+++ forage-android/src/test/java/com/joinforage/forage/android/pos/PosRefundPaymentRepositoryTest.kt
@@ -1,6 +1,5 @@
 package com.joinforage.forage.android.pos
 
-import com.joinforage.forage.android.VaultType
 import com.joinforage.forage.android.core.telemetry.Log
 import com.joinforage.forage.android.fixtures.givenContentId
 import com.joinforage.forage.android.fixtures.givenEncryptionKey
@@ -37,7 +36,7 @@ class PosRefundPaymentRepositoryTest : MockServerSuite() {
     private lateinit var repository: PosRefundPaymentRepository
 
     private lateinit var mockServiceFactory: MockServiceFactory
-    private val mockVaultSubmitter = MockVaultSubmitter(VaultType.VGS_VAULT_TYPE)
+    private val mockVaultSubmitter = MockVaultSubmitter()
     private val expectedData = MockServiceFactory.ExpectedData
     private lateinit var mockForagePinEditText: ForagePINEditText
